pyproject.toml
[tool.poetry]
name = "fl-dl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "fl_dl"}]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
absl-py = "1.0.0"
argon2-cffi = "21.3.0"
argon2-cffi-bindings = "21.2.0"
astunparse = "1.6.3"
attrs = "21.2.0"
backcall = "0.2.0"
beautifulsoup4 = "4.11.1"
bleach = "5.0.1"
cachetools = "3.1.1"
certifi = "2022.6.15"
cffi="1.15.1"
charset-normalizer="2.1.1"
cloudpickle="2.1.0"
cycler="0.11.0"
debugpy="1.6.3"
decorator="5.1.1"
defusedxml="0.7.1"
dill="0.3.5.1"
dm-tree="0.1.7"
entrypoints="0.4"
etils="0.7.1"
farmhashpy="0.4.0"
fastjsonschema="2.16.1"
flatbuffers="2.0.7"
fonttools="4.37.1"
gast="0.4.0"
google-auth="2.11.0"
google-auth-oauthlib="0.4.6"
google-pasta="0.2.0"
googleapis-common-protos="1.56.4"
grpcio="1.37.1"
h5py="3.7.0"
idna="3.3"
importlib-metadata="4.12.0"
importlib-resources="5.9.0"
ipykernel="6.15.2"
ipython="7.34.0"
ipython-genutils="0.2.0"
ipywidgets="8.0.1"
jax="0.2.28"
jaxlib="0.1.76"
jedi="0.18.1"
jinja2="3.1.2"
joblib="1.1.0"
jsonschema="4.15.0"
jupyter="1.0.0"
jupyter-client="7.3.5"
jupyter-console="6.4.4"
jupyter-core="4.11.1"
jupyterlab-pygments="0.2.2"
jupyterlab-widgets="3.0.2"
keras="2.7.0"
keras-nightly="2.11.0.dev2022090207"
keras-preprocessing="1.1.2"
kiwisolver="1.4.4"
libclang="14.0.6"
lxml="4.9.1"
markdown="3.4.1"
markupsafe="2.1.1"
matplotlib="3.5.3"
matplotlib-inline="0.1.6"
mistune="2.0.4"
mpmath="1.2.1"
nbclient="0.6.7"
nbconvert="7.0.0"
nbformat="5.4.0"
nest-asyncio="1.5.5"
notebook="6.4.12"
numpy="1.21.6"
oauthlib="3.2.0"
opt-einsum="3.3.0"
packaging="21.3"
pandocfilters="1.5.0"
parso="0.8.3"
pexpect="4.8.0"
pickleshare="0.7.5"
pillow="9.2.0"
pip="22.1.2"
pkgutil-resolve-name="1.3.10"
portpicker="1.3.9"
prometheus-client="0.14.1"
promise="2.3"
prompt-toolkit="3.0.31"
protobuf="3.19.4"
psutil="5.9.1"
ptyprocess="0.7.0"
pyasn1="0.4.8"
pyasn1-modules="0.2.8"
pycparser="2.21"
pygments="2.13.0"
pyparsing="3.0.9"
pyrsistent="0.18.1"
python-dateutil="2.8.2"
pyzmq="23.2.1"
qtconsole="5.3.2"
qtpy="2.2.0"
requests="2.28.1"
requests-oauthlib="1.3.1"
rsa="4.9"
scikit-learn="1.0.2"
scipy="1.7.3"
semantic-version="2.8.5"
send2trash="1.8.0"
setuptools="63.4.1"
six="1.16.0"
sklearn="0.0"
soupsieve="2.3.2.post1"
tb-nightly="2.11.0a20220901"
tensorboard="2.10.0"
tensorboard-data-server="0.6.1"
tensorboard-plugin-wit="1.8.1"
tensorflow="2.7.0"
tensorflow-datasets="4.6.0"
tensorflow-estimator="2.7.0"
tensorflow-federated-nightly="0.19.0.dev20220218"
tensorflow-io-gcs-filesystem="0.26.0"
tensorflow-metadata="1.10.0"
tensorflow-model-optimization="0.7.3"
tensorflow-privacy="0.7.3"
tensorflow-probability="0.17.0"
termcolor="1.1.0"
terminado="0.15.0"
tf-estimator-nightly="2.11.0.dev2022090208"
tf-nightly="2.11.0.dev20220901"
threadpoolctl="3.1.0"
tinycss2="1.1.1"
toml="0.10.2"
tornado="6.2"
tqdm="4.28.1"
traitlets="5.3.0"
typing-extensions="4.3.0"
urllib3="1.26.12"
wcwidth="0.2.5"
webencodings="0.5.1"
werkzeug="2.2.2"
wheel="0.37.1"
widgetsnbextension="4.0.2"
wrapt="1.14.1"
zipp="3.8.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.15.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
